+++ examples/workflows/aws_privatelink/main.tf
@@ -37,10 +37,10 @@ variable "storage_gib" {
   default  = 15
 }
 
-variable "machine_type" {
-  type     = string
+variable "vcpu_count" {
+  type     = number
   nullable = false
-  default  = "m5.large"
+  default  = 4
 }
 
 variable "subnets" {
@@ -88,8 +88,8 @@ resource "cockroach_cluster" "example" {
   name           = var.cluster_name
   cloud_provider = "AWS"
   dedicated = {
-    storage_gib  = var.storage_gib
-    machine_type = var.machine_type
+    storage_gib      = var.storage_gib
+    num_virtual_cpus = var.vcpu_count
   }
   regions = [{
     name       = var.aws_region