+++ examples/workflows/cockroach_client_ca_cert/main.tf
@@ -27,10 +27,10 @@ variable "storage_gib" {
   default  = 15
 }
 
-variable "machine_type" {
-  type     = string
+variable "vcpu_count" {
+  type     = number
   nullable = false
-  default  = "n2-standard-2"
+  default  = 4
 }
 
 terraform {
@@ -48,8 +48,8 @@ resource "cockroach_cluster" "example" {
   name           = var.cluster_name
   cloud_provider = var.cloud_provider
   dedicated = {
-    storage_gib  = var.storage_gib
-    machine_type = var.machine_type
+    storage_gib      = var.storage_gib
+    num_virtual_cpus = var.vcpu_count
   }
   regions = [
     for r in var.cloud_provider_regions : {