+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Changed
+
+- Remove `machine_type` from examples and clear up in docs that this field
+  requires a feature flag and is not recommended for use.
+
 ### Fixed
 
 - Update docs to reflect that maintenance windows are for Advanced clusters